pyproject.toml
[tool.uv.workspace]
members = ["futuresearch-mcp"]

[tool.uv.sources]
futuresearch = { workspace = true }

[project]
name = "futuresearch"
version = "0.5.1"
description = "A researcher for every row. Forecast, score, classify, or research entire datasets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs>=22.2.0",
    "httpx>=0.20.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dateutil>=2.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/futuresearch"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
case-studies = [
    "ipykernel>=7.1.0",
    "nbconvert>=7.16.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
]
dev = [
    "openapi-python-client>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "basedpyright>=1.22.0",
    "ruff>=0.9.9",
    "jsonschema>=4.0.0",
    "pytest-xdist>=3.8.0",
]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
exclude = ["src/futuresearch/generated"]
typeCheckingMode = "standard"

[tool.ruff]
include = ["src/**/*.py", "tests/**/*.py"]
exclude = [".venv", "src/futuresearch/generated/**"]
target-version = "py312"

[tool.ruff.lint]
select = ["F", "E", "W", "PL", "RUF", "ARG", "F401", "UP"]
ignore = [
    "E501",    # line too long
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
    "PLR2004", # Magic value used in comparison
    "PLR0913", # Too many arguments
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "PLR0911", # Too many return statements
]
fixable = ["ALL"]
unfixable = []
extend-select = ["I"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?)|[a-zA-Z0-9_]+_)$"
exclude = [".venv", "src/futuresearch/generated/**"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.format]
exclude = ["src/futuresearch/generated/**"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib", "-m", "not integration"]
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (require FUTURESEARCH_API_KEY)",
]
asyncio_mode = "auto"
